requirements.txt
faiss-cpu==1.9.0
langchain-community==0.2.16
langchain-google-genai==1.0.10
langchain-openai==0.1.24
langchain==0.2.16
opentelemetry-distro==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp==1.27.0
python-dotenv==1.0.1
streamlit-chat==0.1.1
streamlit==1.38.0
watchdog==5.0.2
